# Conflicts:
#	Sources/Internal/Platform/TemplateWin32/DPIHelperWin32.cpp


Former-commit-id: aca89cf6aa7fac871f1383b3d1b08a27fa37b8e5

+++ Sources/Internal/Platform/TemplateAndroid/AndroidCrashReport.cpp
@@ -99,7 +99,7 @@ void JniCrashReporter::ThrowJavaExpetion(const Vector<CrashStep>& chashSteps)
     jintArray jFileLineArray = env->NewIntArray(chashSteps.size());
 
     int* fileLines = new int[chashSteps.size()];
-    for (uint i = 0; i < chashSteps.size(); ++i)
+    for (size_t i = 0; i < chashSteps.size(); ++i)
     {
         env->SetObjectArrayElement(jModuleArray, i, env->NewStringUTF(chashSteps[i].module));
         env->SetObjectArrayElement(jFunctionArray, i, env->NewStringUTF(chashSteps[i].function));
